Added support for SSL certificate locations
Fixes an issue that happened on some Ruby 1.9 environments, where the
SSL CA certs could not be found. Instead of assuming a location for
them, we allow this location to be specified view the gem
configuration. This is much safer than the previous Ruby standard of not
verifying the certificate (VERIFY_NONE), and does not presume any
installation location for the certs on the client.